requirements.txt
requests==2.31.0
python-dotenv==1.0.0
groq==0.4.2
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2 